requirements.txt
alabaster==0.7.12
altair==4.1.0
applaunchservices==0.2.1
appnope==0.1.0
argh==0.26.2
asn1crypto==1.3.0
astor==0.8.1
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
autograd==1.3
autograd-gamma==0.4.1
autopep8==1.4.4
Babel==2.8.0
backcall==0.1.0
base58==2.0.0
biopython==1.76
bleach==3.1.0
blinker==1.4
boto3==1.13.11
botocore==1.16.11
cachetools==4.1.0
certifi==2020.4.5.1
cffi==1.13.2
chardet==3.0.4
chart-studio==1.0.0
click==7.1.2
cloudpickle==1.2.2
colorlover==0.3.0
conda==4.8.3
conda-package-handling==1.6.0
cryptography==2.8
cufflinks==0.17.0
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
diff-match-patch==20181111
docutils==0.16
entrypoints==0.3
enum-compat==0.0.3
fastcache==1.1.0
flake8==3.7.9
future==0.18.2
gmpy2==2.0.8
idna==2.8
imagesize==1.2.0
importlib-metadata==1.5.0
inflect==4.1.0
intervaltree==3.0.2
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jaraco.itertools==5.0.0
jedi==0.14.1
Jinja2==2.11.1
jmespath==0.10.0
joblib==0.14.1
json5==0.9.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
jupyterlab==1.2.6
jupyterlab-server==1.0.6
keyring==21.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lifelines==0.23.8
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mistune==0.8.4
more-itertools==8.2.0
mpmath==1.1.0
nbconvert==5.6.1
nbformat==5.0.4
neatbio==0.0.2
notebook==6.0.3
numpy==1.18.1
numpydoc==0.9.2
olefile==0.46
packaging==20.1
pandas==1.0.0
pandocfilters==1.4.2
parso==0.6.0
pathtools==0.1.2
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
plotly==4.4.1
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.12.0
psutil==5.6.7
ptyprocess==0.6.0
py==1.8.1
PyBioMed==1.0
pycairo==1.19.1
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pydeck==0.3.1
pydocstyle==4.0.1
pyflakes==2.1.1
Pygments==2.5.2
pylint==2.4.4
Pympler==0.8
pyobjc-core==6.1
pyobjc-framework-Cocoa==6.1
pyobjc-framework-FSEvents==6.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.15.7
PySocks==1.7.1
pytest==5.3.5
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.7
pytz==2019.3
PyYAML==5.3
pyzmq==18.1.1
QDarkStyle==2.8
QtAwesome==0.6.1
qtconsole==4.6.0
QtPy==1.9.0
requests==2.22.0
retrying==1.3.3
rope==0.16.0
Rtree==0.9.3
ruamel-yaml==0.15.87
s3transfer==0.3.3
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
six==1.14.0
snowballstemmer==2.0.0
sortedcontainers==2.1.0
SpeechRecognition==3.6.3
Sphinx==2.3.1
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
spyder==4.0.0
spyder-kernels==1.8.1
statsmodels==0.11.0
streamlit==0.59.0
sympy==1.5.1
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
toolz==0.10.0
tornado==6.0.4
tqdm==4.42.0
traitlets==4.3.3
tzlocal==2.1
ujson==1.35
urllib3==1.25.8
validators==0.15.0
watchdog==0.10.1
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.2
wurlitzer==2.0.0
yapf==0.28.0
zipp==2.1.0
